+++ src/colors/index.ts
@@ -1,8 +1,4 @@
-export {
-  PRIMARY_COLORSET,
-  SECONDARY_COLORSET,
-  FULL_COLORSET,
-} from './colorsets'
+export { PRIMARY_COLORSET, SECONDARY_COLORSET, FULL_COLORSET } from './colorsets'
 
 export * from './colors'
 export * from './color.class'