+++ src/components/buttons/Button/Button.tsx
@@ -1,23 +1,14 @@
 /** @jsx jsx */
 import React, { useContext, Fragment, SyntheticEvent } from 'react'
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
-import {
-  findIconDefinition,
-  IconName,
-  IconPrefix,
-} from '@fortawesome/fontawesome-svg-core'
+import { findIconDefinition, IconName, IconPrefix } from '@fortawesome/fontawesome-svg-core'
 import { jsx } from '@emotion/core'
 
 import { ButtonContainer } from '../ButtonContainer'
 import { ThemeContext } from '../../contexts/ThemeContext'
 import { Color } from '../../../colors'
 
-import {
-  prepareBaseState,
-  prepareFilledState,
-  iconOnlyState,
-  rawLinkState,
-} from './styles'
+import { prepareBaseState, prepareFilledState, iconOnlyState, rawLinkState } from './styles'
 
 export const Button: React.FC<ButtonProps> = props => {
   const theme = useContext(ThemeContext)