+++ src/components/layout/ThematicSection/ThematicSection.tsx
@@ -3,30 +3,17 @@ import React from 'react'
 import { ReactElementLike } from 'prop-types'
 import { jsx } from '@emotion/core'
 
-import {
-  FerpectionBadgeIcon,
-  FerpectionBadgeIconProps,
-} from '../../iconography/FerpectionBadgeIcon'
+import { FerpectionBadgeIcon, FerpectionBadgeIconProps } from '../../iconography/FerpectionBadgeIcon'
 import { Color } from '../../../colors'
 
-import {
-  containerStyles,
-  headerStyles,
-  titleStyles,
-  sectionStyles,
-  titleTextStyles,
-  logoStyles,
-} from './styles'
+import { containerStyles, headerStyles, titleStyles, sectionStyles, titleTextStyles, logoStyles } from './styles'
 
 export const ThematicSection: React.FC<ThematicSectionProps> = props => (
   <div css={[containerStyles]}>
     <header css={[headerStyles]}>
       <div css={[titleStyles]}>
         <div css={[logoStyles]}>
-          <FerpectionBadgeIcon
-            icon={props.badgeIcon}
-            borderColor={props.badgeColor}
-          />
+          <FerpectionBadgeIcon icon={props.badgeIcon} borderColor={props.badgeColor} />
         </div>
         <h2 css={[titleTextStyles]}>{props.title}</h2>
       </div>